requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
faiss-cpu
sentence-transformers
gradio
pypdf
python-dotenv
